[department-of-veterans-affairs/orchid#31] Sorts topics in ABC order and extracts medical centers into own file
